requirements.txt
Flask==1.1.1
Werkzeug==0.16.0
Flask_Dance==2.2.0
Flask-Mail==0.9.1
PyYAML==5.1.2
python-dateutil==2.8.1
requests==2.22.0
cryptography==2.7
packaging==19.2
sqlalchemy==1.3.12
SQLAlchemy-Utils==0.36.1
flask-sqlalchemy==2.4.1
Flask-Migrate==2.5.2
Flask-Alembic==2.0.1
alembic==0.9.9
Flask-Script==2.0.6
visitor==0.1.3
tzlocal==2.0.0
PyMySQL==0.9.3
python-swiftclient==3.9.0
python-keystoneclient==3.22.0
pycryptodomex==3.9.7
hvac==0.9.6
MarkupSafe==1.1.1